+++ attributemap/amazon2name.php
@@ -1,0 +1,7 @@
+<?php
+
+$attributemap = array(
+    'amazon.name' => ['cn', 'displayName'],
+    'amazon.email' => 'mail',
+    'amazon.user_id' => 'uid',
+);